adding checkpoints is a separate component
