+++ Explorer/src/app/feature-modules/tour-authoring/add-tour-checkpoints/add-tour-checkpoints.component.css
@@ -1,0 +1,299 @@
+/* Page Layout */
+.page-container {
+    display: flex;
+    gap: 2rem;
+    padding: 2rem;
+    max-width: 1400px;
+    margin: 0 auto;
+  }
+  
+  /* Forms Section */
+  .forms-section {
+    flex: 1;
+    min-width: 0;
+  }
+  
+  .form-container {
+    background-color: var(--white);
+    border-radius: 8px;
+    padding: 1.5rem;
+    margin-bottom: 2rem;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .custom-form {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+  
+  /* Form Elements */
+  .form-group {
+    margin-bottom: 1rem;
+  }
+  
+  .form-label {
+    display: block;
+    color: var(--text-light);
+    margin-bottom: 0.5rem;
+    font-weight: 500;
+  }
+  
+  .input-field,
+  .select-field {
+    width: 100%;
+    padding: 0.75rem;
+    border: 1px solid var(--light-brown);
+    border-radius: 4px;
+    color: var(--text-dark);
+    transition: border-color 0.3s ease;
+    background-color: var(--white);
+  }
+  
+  .input-field:focus,
+  .select-field:focus {
+    outline: none;
+    border-color: var(--primary-brown);
+    box-shadow: 0 0 0 2px rgba(139, 115, 85, 0.1);
+  }
+  
+  .textarea-field {
+    width: 100%;
+    min-height: 120px;
+    padding: 0.75rem;
+    border: 1px solid var(--light-brown);
+    border-radius: 4px;
+    resize: vertical;
+  }
+  
+  .coordinates-group {
+    display: flex;
+    gap: 1rem;
+  }
+  
+  .coordinates-group .form-group {
+    flex: 1;
+  }
+  
+  /* File Input */
+  .file-input {
+    width: 100%;
+    padding: 0.5rem 0;
+  }
+  
+  /* Image Preview */
+  .image-preview {
+    margin: 1rem 0;
+  }
+  
+  .preview-title {
+    color: var(--text-dark);
+    margin-bottom: 0.5rem;
+  }
+  
+  .preview-image {
+    max-width: 200px;
+    border-radius: 4px;
+    border: 1px solid var(--light-brown);
+  }
+  
+  /* Map Container */
+  .map-container {
+    margin: 1rem 0;
+    border-radius: 8px;
+    overflow: hidden;
+    border: 1px solid var(--light-brown);
+  }
+  
+  /* Table Section */
+  .table-section {
+    flex: 1;
+    min-width: 0;
+  }
+  
+  .table-container {
+    background-color: var(--white);
+    border-radius: 8px;
+    padding: 1rem;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    overflow-x: auto;
+  }
+  
+  .custom-table {
+    width: 100%;
+    border-collapse: collapse;
+  }
+  
+  .custom-table th {
+    background-color: var(--lighter-brown);
+    color: var(--dark-brown);
+    font-weight: 600;
+    text-align: left;
+    padding: 1rem;
+  }
+  
+  .custom-table td {
+    padding: 1rem;
+    border-bottom: 1px solid var(--lighter-brown);
+    color: var(--text-light);
+  }
+  
+  .custom-table tr:hover {
+    background-color: rgba(210, 180, 140, 0.1);
+  }
+  
+  .checkpoint-image {
+    max-width: 150px;
+    max-height: 150px;
+    border-radius: 4px;
+    object-fit: cover;
+  }
+  
+  /* Submit Button */
+  .form-actions {
+    margin-top: 1rem;
+    text-align: right;
+  }
+  
+  .submit-button {
+    background-color: var(--primary-brown);
+    color: var(--white);
+    padding: 0.75rem 1.5rem;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-weight: 500;
+    transition: all 0.3s ease;
+  }
+  
+  .submit-button:hover {
+    background-color: var(--dark-brown);
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  }
+  
+  .submit-button:disabled {
+    background-color: var(--text-light);
+    cursor: not-allowed;
+  }
+  
+  /* Responsive Design */
+  @media (max-width: 1024px) {
+    .page-container {
+      flex-direction: column;
+    }
+  
+    .coordinates-group {
+      flex-direction: column;
+      gap: 0;
+    }
+  }
+  
+  @media (max-width: 640px) {
+    .page-container {
+      padding: 1rem;
+    }
+  
+    .form-container,
+    .table-container {
+      padding: 1rem;
+    }
+  
+    .custom-table th,
+    .custom-table td {
+      padding: 0.75rem;
+    }
+  }
+  
+  /* Add to your existing component styles */
+  
+  .image-preview {
+    margin-top: 1rem;
+    padding: 1rem;
+    border-radius: 8px;
+    background-color: var(--white);
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    animation: fadeIn 0.3s ease-in-out;
+  }
+  
+  .preview-title {
+    color: var(--dark-brown);
+    font-size: 1rem;
+    font-weight: 500;
+    margin-bottom: 1rem;
+    padding-bottom: 0.5rem;
+    border-bottom: 2px solid var(--lighter-brown);
+  }
+  
+  .preview-image {
+    width: 100%;
+    max-height: 300px;
+    object-fit: contain;
+    border-radius: 4px;
+    display: block;
+    margin: 0 auto;
+    border: 1px solid var(--lighter-brown);
+    background-color: #f8f8f8;
+    transition: transform 0.3s ease;
+  }
+  
+  .preview-image:hover {
+    transform: scale(1.02);
+    cursor: pointer;
+  }
+  
+  @keyframes fadeIn {
+    from {
+      opacity: 0;
+      transform: translateY(10px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+  
+  /* Media query for responsive design */
+  @media (min-width: 768px) {
+    .image-preview {
+      margin-top: 1.5rem;
+      padding: 1.5rem;
+    }
+  
+    .preview-image {
+      max-height: 400px;
+    }
+  }
+  
+  /* Add these styles to your existing CSS file */
+  
+  .file-input-container {
+    margin: 1rem 0;
+  }
+  
+  .file-label {
+    display: block;
+    width: 100%;
+    padding: 1rem;
+    text-align: center;
+    border: 2px dashed var(--light-brown);
+    border-radius: 4px;
+    color: var(--text-dark);
+    cursor: pointer;
+    transition: all 0.3s ease;
+    background-color: transparent;
+  }
+  
+  .file-label:hover {
+    border-color: var(--primary-brown);
+    color: var(--primary-brown);
+  }
+  
+  .file-input {
+    display: none;
+  }
+  
+  /* Optional: Add this if you want to style the text inside */
+  .upload-text {
+    font-size: 1rem;
+    color: var(--dark-brown);
+  }
