+++ Explorer/src/app/infrastructure/routing/app-routing.module.ts
@@ -32,6 +32,7 @@ import { SendMessageComponent } from 'src/app/feature-modules/stakeholders/send-
 import { PurchasedToursComponent } from 'src/app/feature-modules/tour-execution/purchased-tours/purchased-tours.component';
 import { TourPreferenceComponent } from 'src/app/feature-modules/tour-authoring/tour-preference/tour-preference.component';
 import { FaqComponent } from 'src/app/feature-modules/administration/faq/faq.component';
+import { AddTourCheckpointsComponent } from 'src/app/feature-modules/tour-authoring/add-tour-checkpoints/add-tour-checkpoints.component';
 
 const routes: Routes = [
   {path: '', component: HomeComponent},
@@ -71,6 +72,7 @@ const routes: Routes = [
   {path: 'shopping-cart/purchasedTours', component: PurchasedToursComponent, canActivate: [AuthGuard]},
   {path: 'tourPreference', component: TourPreferenceComponent, canActivate: [AuthGuard]},
   {path: 'faq', component: FaqComponent, canActivate: [AuthGuard]},
+  {path: 'tour/:id/checkpoints', component: AddTourCheckpointsComponent, canActivate: [AuthGuard]},
 ];
 
 @NgModule({
